requirements.txt
asgiref==3.4.0
beautifulsoup4==4.9.3
Django==3.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
